complete featured and restructure image and update global css
