+++ pages/api/components/Featured/ProjectItem.jsx
@@ -1,0 +1,17 @@
+import styles from "../../../../styles/Featured/ProjectItem.module.css";
+
+const ProjectItem = ({ item }) => {
+  return (
+    <div>
+      <div className={styles["img-container"]}>
+        <img src={item.image} alt={item.title} />
+      </div>
+      <div className={styles.text}>
+        <p className={styles.p}>{item.title}</p>
+        <h3>{item.subtitle}</h3>
+      </div>
+    </div>
+  );
+};
+
+export default ProjectItem;
