+++ styles/Featured/Featured.module.css
@@ -31,6 +31,7 @@
 }
 
 .lower-span {
+  padding: 0 20px;
   display: flex;
   justify-content: space-between;
   gap: 40px;
