+++ styles/Featured/ProjectItem.module.css
@@ -1,0 +1,21 @@
+.img-container {
+  width: 340px;
+  height: 340px;
+  overflow: hidden;
+}
+
+.img-container img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.text {
+  width: 316px;
+  padding: 12px;
+  text-align: left;
+}
+
+.p {
+  margin-bottom: 4px;
+}
