+++ styles/globals.css
@@ -18,6 +18,13 @@ h2 {
   line-height: 48px;
 }
 
+h3 {
+  font-family: Roboto;
+  font-size: 20px;
+  font-weight: 500;
+  line-height: 28px;
+}
+
 p {
   color: #000;
   font-family: Roboto;
